delete ReqDtoB class
